completed rules section

+++ src/popup/pages/rules/RuleGroup.tsx
@@ -1,12 +1,19 @@
+import { yupResolver } from '@hookform/resolvers/yup';
 import DeleteIcon from '@mui/icons-material/Delete';
 import EditIcon from '@mui/icons-material/Edit';
 import { Switch, Tooltip } from '@mui/material';
 import React from 'react';
+import { Resolver, useFieldArray, useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
+import { ConditionValues, RuleType } from '../../../types';
 import Rule from '../../../utils/Rule';
+import formSchema from '../../../utils/formSchema';
 import Circle from '../../components/Circle';
+import ConditionGroup from '../../components/ConditionGroup';
+import GroupBuilder from '../../components/GroupBuilder';
 import RowGroup from '../../components/RowGroupParent';
 import { useAlertProvider } from '../../provider/AlertProvider';
+import { useModal } from '../../provider/ModalProvider';
 import { usePopupStatus } from '../../provider/PopupStatusProvider';
 
 interface Props {
@@ -18,6 +25,7 @@ export default function RuleGroup({ rule, updateRules }: Props) {
    const navigate = useNavigate();
    const isPopup = usePopupStatus();
    const { setAlertSettings } = useAlertProvider();
+   const { getOutput } = useModal();
 
    /**
     * Handles toggling active status of rule
@@ -34,8 +42,15 @@ export default function RuleGroup({ rule, updateRules }: Props) {
     */
    const handleDeleteRule = async () => {
       try {
-         await rule.delete();
-         await updateRules();
+         const output = await getOutput({
+            title: 'Delete rule?',
+            type: 'confirmation',
+            body: 'This will permanently delete this rule from your Chrome storage.',
+         });
+         if (output) {
+            await rule.delete();
+            await updateRules();
+         }
       } catch (err) {
          console.error(err);
          setAlertSettings('error', 'Something went wrong.');
@@ -48,6 +63,36 @@ export default function RuleGroup({ rule, updateRules }: Props) {
    const handleEditRule = async () => {
       navigate('edit', { state: { ruleId: rule.id } });
    };
+
+   /**
+    * Create form for rendering ConditionGroup component
+    */
+   const { control } = useForm<RuleType | ConditionValues>({
+      defaultValues: { conditionGroups: rule.conditionGroups },
+      resolver: yupResolver(formSchema) as unknown as Resolver<
+         ConditionValues | Partial<RuleType>
+      >,
+   });
+
+   /**
+    * Create field array from rule.conditionGroups.groups
+    */
+   const { fields } = useFieldArray({
+      control,
+      name: 'conditionGroups.groups',
+   });
+
+   /**
+    * Render a component for each field in field array for use in GroupBuilder component
+    */
+   const renderedGroups = fields.map((field, index) => (
+      <ConditionGroup
+         key={field.id}
+         index={index}
+         control={control}
+         read_only={true}
+      />
+   ));
 
    return (
       <RowGroup
@@ -79,21 +124,10 @@ export default function RuleGroup({ rule, updateRules }: Props) {
          title={rule.title}
          secondary={rule.formatActionText()}
       >
-         {/* {rule.conditions.map((condition: ConditionType) => (
-            <Row
-               key={condition.query}
-               id={condition.id}
-               isChild={true}
-               title={Rule.formatConditionText(condition)}
-               AffixIcon={
-                  <Tooltip title='Remove condition from rule'>
-                     <CloseIcon fontSize='small' />
-                  </Tooltip>
-               }
-               affixAction={() => handleDeleteCondition(condition.id)}
-            />
-         ))} */}
-         <></>
+         <GroupBuilder
+            childrenArr={renderedGroups}
+            label={rule.conditionGroups.all_required ? 'AND' : 'OR'}
+         />
       </RowGroup>
    );
 }
